# Minor changes to improve compilation speed