This crate is small and compiles quickly, but it's also used *everywhere*, so trying to improve the compilation speed is worth it.

This PR has a number of optimizations; each commit is independent, so feel free to cherry-pick and take whichever changes you want.  This PR preserves the current MSRV of 1.13.

Changes:
- Add `rust-version = "1.13"` to the `Cargo.toml`, to partially document the current MSRV
- Change internal methods that use `AsRef<Path>` to just take `&Path`
- Use loops instead of iterator combinators in parsing functions
- Make the `debug!` macro generate no code when not in use
- Use a `Vec` instead of a `HashMap` when counting physical cores; using the `HashMap` requires compiling a lot more code.  (The `Vec` should also be faster.)

My non-scientific benchmarks of a clean build of the library (`libc` + `num_cpus`)
- in release mode: 0.8s -> 0.7s (-13%)
- in debug mode: 0.63s -> 0.60s (-5%)

Measuring the amount of code passed to LLVM, using [cargo-llvm-lines](https://github.com/dtolnay/cargo-llvm-lines):  
`cargo llvm-lines --lib -p num_cpus` (with `--release` for release builds)
- in release mode: 16325 -> 13039 (-20%)
- in debug mode: 24263 -> 18781 (-22%)