@Aeledfyr 2024-02-25
Improved benchmarks, only compiling `num_cpus` by directly invoking rustc (with the command given by cargo's `--verbose):
- release build: from 364.4 ms ±  7.6 ms to 270.4 ms ±  7.8 ms (-25%)
- debug build: from 70.1 ms ±  3.0 ms to 60.4 ms ±   2.7 ms (-13%)