@Aeledfyr 2024-02-26
It looks like most of the CI failures are due to an updated Rust toolchain:
- Nightly tests fail to compile because of a new warning (and `#[deny(warnings)]`)
- Cross compiling to some architectures fails because of a missing `rust-std` components

The only test that actually fails is `test-cgroups`, which fails in the initial (non-cgroups) test, because it expects to be run on a machine with 2 cores, but the CI now runs it on a 4-core machine.  I'm not 100% sure that that's what is happening, but it may be worth rerunning CI on the main branch to see if that's the issue.